requirements.txt
groq
aiohttp
requests
urllib3
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
